Added jumping to ball and props to the scene.
After adding the ability for the 'player' ball to jump, I was then
challenged to figure out how to stop the ability to jump over and over
again, essentially becoming like a 3D version of the smartphone app
"Flappy Bird". I stopped this by coming up with a way to know whether or
not the ball is in the air or on the ground. So using boolean values I
made it so the ball can only jump if it's on the ground. Fun stuff!